+++ RELEASE-NOTES.md
@@ -1,5 +1,14 @@
 ## RELEASE NOTES
 
+### Version 2.49.2 - May 21 2019
+**RDM-876** Hide print button when not configured  
+
+### Version 2.48.0 - May 7 2019
+**RDM-4575** search filters definition service
+
+### Version 2.47.0 - May 7 2019
+**RDM-4574** Create Case Filters componant retrieve jurisdiction internally 
+
 ### Version 2.46.0 - May 1 2019
 **RDM-3496** Case View of Nested Complex Types - Collection of complex types within complex types